Make Widget IDs backed by an Int
This is wildly more efficient in JS. I started looking at expect/actual-ing this storage type to be Double on JS and Long on other platforms but an Int means we can create 1000 widgets per second for 49 days straight without overflow. Seems fine.

+++ redwood-protocol/src/commonMain/kotlin/app/cash/redwood/protocol/values.kt
@@ -21,9 +21,9 @@ import kotlinx.serialization.Serializable
 /** Identifies a widget instance. */
 @JvmInline
 @Serializable
-public value class Id(public val value: Long) {
+public value class Id(public val value: Int) {
   public companion object {
-    public val Root: Id = Id(0L)
+    public val Root: Id = Id(0)
   }
 }
